requirements.txt
colorama==0.4.6
iniconfig==2.3.0
mysql-connector-python==9.6.0
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.2
